backend/requirements.txt
# --- Core Data & ML ---
numpy
pandas
nltk
sentence-transformers
huggingface-hub
openai
vaderSentiment==3.3.2

# --- Web Scraping & APIs ---
snscrape==0.7.0.20230622
praw==7.8.1
requests
beautifulsoup4
google-api-python-client
youtube-transcript-api
google-search-results
selenium
webdriver-manager
ntscraper

# --- Database ---
sqlalchemy
mysql-connector-python
psycopg2-binary

# --- Web Framework ---
fastapi
uvicorn[standard]

# --- Utilities ---
python-dotenv
tqdm
markdown2

# --- Plotting ---
matplotlib
seaborn
google-auth